requirements.txt
flask
telethon
python-dotenv
gunicorn
redis
